+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/JiraLaunchTimeouts.kt
@@ -1,0 +1,9 @@
+package com.atlassian.performance.tools.infrastructure.api.jira
+
+import java.time.Duration
+
+data class JiraLaunchTimeouts(
+    val offlineTimeout: Duration,
+    val initTimeout: Duration,
+    val upgradeTimeout: Duration
+)