+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/JiraNodeConfig.kt
@@ -6,14 +6,49 @@ import com.atlassian.performance.tools.infrastructure.api.jvm.jmx.DisabledRemote
 import com.atlassian.performance.tools.infrastructure.api.jvm.jmx.RemoteJmx
 import com.atlassian.performance.tools.infrastructure.api.splunk.DisabledSplunkForwarder
 import com.atlassian.performance.tools.infrastructure.api.splunk.SplunkForwarder
+import java.time.Duration
 
 data class JiraNodeConfig(
-    val name: String = "jira-node",
-    val debug: JvmDebug = DisabledJvmDebug(),
-    val remoteJmx: RemoteJmx = DisabledRemoteJmx(),
-    val jvmArgs: JiraJvmArgs = JiraJvmArgs(),
-    val splunkForwarder: SplunkForwarder = DisabledSplunkForwarder()
+    val name: String,
+    val debug: JvmDebug,
+    val remoteJmx: RemoteJmx,
+    val jvmArgs: JiraJvmArgs,
+    val splunkForwarder: SplunkForwarder,
+    val launchTimeouts: JiraLaunchTimeouts
 ) {
+    constructor(
+        name: String,
+        jvmArgs: JiraJvmArgs,
+        launchTimeouts: JiraLaunchTimeouts
+    ) : this(
+        name = name,
+        debug = DisabledJvmDebug(),
+        remoteJmx = DisabledRemoteJmx(),
+        jvmArgs = jvmArgs,
+        splunkForwarder = DisabledSplunkForwarder(),
+        launchTimeouts = launchTimeouts
+    )
+
+    @Deprecated(message = "Use the other constructor")
+    constructor(
+        name: String = "jira-node",
+        debug: JvmDebug = DisabledJvmDebug(),
+        remoteJmx: RemoteJmx = DisabledRemoteJmx(),
+        jvmArgs: JiraJvmArgs = JiraJvmArgs(),
+        splunkForwarder: SplunkForwarder = DisabledSplunkForwarder()
+    ) : this(
+        name = name,
+        debug = debug,
+        remoteJmx = remoteJmx,
+        jvmArgs = jvmArgs,
+        splunkForwarder = splunkForwarder,
+        launchTimeouts = JiraLaunchTimeouts(
+            offlineTimeout = Duration.ofMinutes(8),
+            initTimeout = Duration.ofMinutes(4),
+            upgradeTimeout = Duration.ofMinutes(8)
+        )
+    )
+
     fun clone(
         times: Int
     ): List<JiraNodeConfig> = (1..times).map { copy(name = "$name-$it") }